Rename ResourceMetadata/NameConstraints to ArmResourceMetadata/ArmResourceNameConstraints and refactor extensions to C# 14 syntax
